--- a/go.mod
+++ b/go.mod
@@ -7,6 +7,7 @@ replace github.com/iden3/go-circom-prover-verifier => ./
 require (
 	github.com/ethereum/go-ethereum v1.9.12
 	github.com/iden3/go-circom-prover-verifier v0.0.0-00010101000000-000000000000
+	github.com/iden3/go-iden3-crypto v0.0.4
 	github.com/stretchr/testify v1.5.1
 	golang.org/x/crypto v0.0.0-20200311171314-f7b00557c8c4
 )