--- a/prover/arithmetic.go
+++ b/prover/arithmetic.go
@@ -3,12 +3,21 @@ package prover
 import (
 	"bytes"
 	"math/big"
+
+	"github.com/iden3/go-iden3-crypto/ff"
 )
 
 func arrayOfZeroes(n int) []*big.Int {
 	var r []*big.Int
 	for i := 0; i < n; i++ {
 		r = append(r, new(big.Int).SetInt64(0))
+	}
+	return r
+}
+func arrayOfZeroesE(n int) []*ff.Element {
+	var r []*ff.Element
+	for i := 0; i < n; i++ {
+		r = append(r, ff.NewElement())
 	}
 	return r
 }
@@ -75,11 +84,32 @@ func polynomialSub(a, b []*big.Int) []*big.Int {
 	return r
 }
 
+func polynomialSubE(a, b []*ff.Element) []*ff.Element {
+	r := arrayOfZeroesE(max(len(a), len(b)))
+	for i := 0; i < len(a); i++ {
+		r[i].Add(r[i], a[i])
+	}
+	for i := 0; i < len(b); i++ {
+		r[i].Sub(r[i], b[i])
+	}
+	return r
+}
+
 func polynomialMul(a, b []*big.Int) []*big.Int {
 	r := arrayOfZeroes(len(a) + len(b) - 1)
 	for i := 0; i < len(a); i++ {
 		for j := 0; j < len(b); j++ {
 			r[i+j] = fAdd(r[i+j], fMul(a[i], b[j]))
+		}
+	}
+	return r
+}
+
+func polynomialMulE(a, b []*ff.Element) []*ff.Element {
+	r := arrayOfZeroesE(len(a) + len(b) - 1)
+	for i := 0; i < len(a); i++ {
+		for j := 0; j < len(b); j++ {
+			r[i+j].Add(r[i+j], ff.NewElement().Mul(a[i], b[j]))
 		}
 	}
 	return r
@@ -99,4 +129,32 @@ func polynomialDiv(a, b []*big.Int) ([]*big.Int, []*big.Int) {
 		rem = aux2[:len(aux2)-1]
 	}
 	return r, rem
+}
+
+func polynomialDivE(a, b []*ff.Element) ([]*ff.Element, []*ff.Element) {
+	// https://en.wikipedia.org/wiki/Division_algorithm
+	r := arrayOfZeroesE(len(a) - len(b) + 1)
+	rem := a
+	for len(rem) >= len(b) {
+		l := ff.NewElement().Div(rem[len(rem)-1], b[len(b)-1])
+		pos := len(rem) - len(b)
+		r[pos] = l
+		aux := arrayOfZeroesE(pos)
+		aux1 := append(aux, l)
+		aux2 := polynomialSubE(rem, polynomialMulE(b, aux1))
+		rem = aux2[:len(aux2)-1]
+	}
+	return r, rem
+}
+
+// once https://github.com/iden3/go-iden3-crypto/pull/22 is merged, use the fucntion from there
+func ElementArrayToBigIntArray(e []*ff.Element) []*big.Int {
+	var o []*big.Int
+	for i := range e {
+		ei := e[i]
+		bi := big.NewInt(0)
+		ei.ToBigIntRegular(bi)
+		o = append(o, bi)
+	}
+	return o
 }